More WIP match code
takes the matching tokens from MATCH_ROW and MATCH_COL, puts them into an easily traversable array and deletes them.  Right now the code only adds the latest matching token, not the first, so there will be holes of one on the board, which doesn´t make sense, so will fix that next :) And of course, matches shouldn´t be added unless there are at least 3 in a row. Step by step..