even more WIP matching code
added function to mark matches (still anything 2 tokens or greater)
added buttons to refresh board, mark, and delete matching tokens, to make testing easier.

have a good way, I think for how to only mark 3 or greater. Commented in code:

to only store matches 3 or larger, after each dud, if we find another match we store in an array within the array and then ignore any subarrays smaller than 3 in size for subsequent actions. So, we need an iterator for every new match we make.